requirements.txt
pillow>=9.0.0
mutagen>=1.45.0
PyPDF2>=3.0.0
